fix(service): service `salt-master` and 'salt-minion' to (re)start last.
When running a high-state on the salt-master to deploy itself, the run fails with an `Authentication error occurred` because the master restarts half way though. When I add the `order: last` to the respective service. states, the high-state runs through. I do not have the time unfortunately to test this through, but at first sight it seems not to lead to problems. Thereby I cannot come up with a reason to have the master or minion restart before all the other states have finished.